requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.2
black==21.7b0
click==8.0.1
Django==3.2.6
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.3
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
wrapt==1.12.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.2
attrs==21.2.0
black==21.7b0
click==8.0.1
Django==3.2.6
djangorestframework==3.12.4
drf-spectacular==0.17.3
inflection==0.5.1
isort==5.9.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
uritemplate==3.0.1
wrapt==1.12.1
